V4.12.1.320 - Back porting changes to MKL devices

+++ net.sourceforge.usbdm.deviceEditor/src/net/sourceforge/usbdm/deviceEditor/information/LongVariable.java
@@ -614,8 +614,8 @@ public class LongVariable extends Variable {
       return fRadix;
    }
 
-   public void setRadix(int radix) {
-      fRadix = radix;
+   public void setRadix(long radix) {
+      fRadix = (int) radix;
    }
 
    @Override
